+++ src/index.ts
@@ -7,6 +7,7 @@ import TraceConfig from './plugins/config/TraceConfig';
 import MetricConfig from './plugins/config/MetricConfig';
 import InvocationConfig from './plugins/config/InvocationConfig';
 import TraceDef from './plugins/config/TraceDef';
+import Utils from './plugins/Utils';
 
 const ThundraWarmup = require('@thundra/warmup');
 
@@ -31,9 +32,12 @@ module.exports = (options: any) => {
         const metricPlugin = MetricPlugin(config.metricConfig);
         config.plugins.push(metricPlugin);
     }
+    const applicationId = process.env.AWS_LAMBDA_LOG_STREAM_NAME ?
+                          process.env.AWS_LAMBDA_LOG_STREAM_NAME.split(']').pop() :
+                          Utils.generateId();
 
     const pluginContext = {
-        applicationId: process.env.AWS_LAMBDA_LOG_STREAM_NAME.split(']').pop(),
+        applicationId,
         applicationProfile: process.env.thundra_applicationProfile ? process.env.thundra_applicationProfile : 'default',
         applicationRegion: process.env.AWS_REGION,
         applicationVersion: process.env.AWS_LAMBDA_FUNCTION_VERSION,